sampling rate in UI now as floats

+++ mobile/src/main/java/es/uni_freiburg/de/cmotion/adapter/SensorAdapter.java
@@ -30,7 +30,7 @@ public class SensorAdapter extends RecyclerView.Adapter implements AbstractDataA
         @Override
         public void onTextChanged(Object tag, String newText) {
             int pos = (int) tag;
-            mCollection.get(pos).setSamplingRate((int) Float.parseFloat(newText));
+            mCollection.get(pos).setSamplingRate(Float.parseFloat(newText));
             notifyItemChanged(pos);
         }
     };
